+++ release-notes.md
@@ -1,2 +1,6 @@
-# Sprint 1.2.4 Lot1.2
-CLI framework foundations.
+# Sprint 1.2.4 - Lot 1.3
+
+## New
+- Added 'info' command
+- Added system information library
+- Added basic integration test